rename idstore to tnodb_mock
